compiler: support schema files with unused imports
Fixes: https://github.com/pelzlpj/capnp-ocaml/issues/8

When compiling a schema file with unused imports, capnpc may emit a
CodeGeneratorRequest which includes an Import structure that has no
corresponding Node structure.  The compiler had expected a Node structure in
order to construct a generated module name; however, it looks like we can get
the information we need directly from the Import structure.