compiler: omit imports resulting in unused modules
Fixes: https://github.com/pelzlpj/capnp-ocaml/issues/6

In the past, capnp-ocaml had unconditionally emitted a module reference as the
result of any import statement.  This has proven to be burdensome due to the
common import of "/capnp/c++.capnp", which is pointless for OCaml code
generation and especially confusing due to the need to mangle its generated
module name.

With this change, we adopt the capnpc-c++ strategy of emitting a module
reference for an import only when the content of that module is actually
referenced by the importing schema.