Disable some warnings in generated files
There's nothing that users of the files can do about them. Example
warnings:

    Warning 32: unused value has_call.
    Warning 27: unused variable x.
    Warning 37: constructor Struct is never used to build values.
    Warning 60: unused module DefaultsCopier_.

This requires PPX and doesn't work with camlp4. I've therefore upped the
minimum OCaml version and disabled camlp4 for the files that don't need
it (which includes all the autogenerated ones).

Note that the tests already didn't work on 4.01:

    # File "testCodecs.ml", line 35, characters 20-55:
    # Error: Unbound module Core_extended.Quickcheck_deprecated