Update to latest core_kernel
Core_kernel.Core_int64 no longer exists. However, it seems we're not
using anything that isn't in the stdlib's version anyway, and reducing
the dependency of the runtime on Core_kernel seems useful.

Dequeue is now called Deque.

Also, remove a gmon.out file, which I assume wasn't supposed to be committed.