Add parameterised reader_t and builder_t types
Add `reader_t` and `builder_t` types to the generated files, plus an
abstract type for each struct. Each struct reader is then defined as
`struct_type reader_t`, and each builder as `struct_type builder_t`.
This makes it possible to define generic functions that work on all
readers or builders, without having to generate a separate function
for each one.

Use this to replace all the `of_pointer` functions with a single
`Reader.of_pointer`.

This also makes interface types abstract, in preparation for allowing
capabilities to be attached to messages. The types were previously
declared as `int option`, but this wasn't exposed or used anywhere.