Share the definitions of reader_t and builder_t
Before, each generated module defined its own types.
This meant that a struct reader from one schema file (e.g. from an
AnyPointer) couldn't be used with another.

This was a problem for the RPC system, which would like to give user
code a request or response struct of the expected type, but this type
is defined in the application's own schema file.

To do this, I have added a phantom type to StructStorage, and defined
StructStorage.reader_t and StructStorage.builder_t types using it.
The generated schema files expose that their reader and builder types
are equal to these.

We don't want to expose the fact that all StructStorage types are really
equal to the compiler, since then it can't detect when the user isn't
following the schema. Instread, StructStorage.t is now private and you
must use the `StructStorage.v` constructor to create them. This prevents
accidents, while not preventing the user from casting between struct
types if they want to. We need to allow this anyway because the
generated modules need to do this, and they have no special access to
the library's internals.