Use polymorphic variants for phantom types
Instead of declaring a load of unique type names at the start of the
file, use polymorphic variants. These don't need to be declared before
use.

e.g.

    type struct_Foo_11668594054516559475
    type reader_t_Foo_11668594054516559475 = struct_Foo_11668594054516559475 reader_t
    type builder_t_Foo_11668594054516559475 = struct_Foo_11668594054516559475 builder_t

    module Foo : sig
      type t = reader_t_Foo_11668594054516559475
      type builder_t = builder_t_Foo_11668594054516559475
      val of_builder : builder_t -> t

becomes

    module Foo : sig
      type struct_t = [`Foo_a1ef2c783fb1ce73]
      type t = struct_t reader_t
      val of_builder : struct_t builder_t -> t

This simplifies imports (all uses of `Obj.magic` have been removed from
the generated code).

It also makes it possible to refer to a type before applying the
generated code's functor. This might be useful for the RPC system, as
it allows it to define a capability as something that takes a request
struct, before applying the RPC schema's functor (that defines what a
request struct is).

I also switched to formatting the node ID as hex, to make it shorter and
to match the format used in the schema files.

Note that there is a slight API change here, because the `$Struct.reader_t`
and `$Struct.builder_t` types have gone (to avoid confusion with the
generic `reader_t` and `builder_t` types). You have to use `struct_t reader_t`
and `struct_t builder_t` instead.

To make writing generic code simpler, there are two new helpers in
StructStorage:

    val reader_of_builder : 'a builder_t -> 'a reader_t
    val message_of_builder : _ builder_t -> rw Message.t

This required some changes to the benchmark code, but the new code is
shorter (`6 files changed, 29 insertions(+), 57 deletions(-)`).