Generate code for methods
Readers and builders for the implicit parameters and results structures
are now generated.

A new Client module contains a sub-module for each interface, with a
sub-module for each method. Each method contains `Params` and `Results`
sub-module aliases (params to a builder, results to a reader) and the
method ID.

A new Service module provides a similar structure for service authors,
but where `Params` is a reader and `Results` is a builder.