Add support for RPC / interfaces
This PR extends the schema compiler to support Cap'n Proto's RPC
system. It does not actually implement an RPC system itself, but
provides the features an RPC provider needs.

There is a new `MakeRPC` functor in the generated schema that takes an
extended version of MessageSig. The original `Make` is still provided -
it is now a convenience wrapper around `MakeRPC` that passes some dummy
functions.

Extending `MessageSig` rather than taking a second argument makes the
code cleaner, and is necessary to avoid an OCaml inlining bug
(https://caml.inria.fr/mantis/view.php?id=7538).

For changes to the generated code, see the updates to the README.adoc file.

For information about the API between the RPC system and the generated
code, see the comments in `src/runtime/rPC.ml`.

There is an RPC implementation available at
<https://github.com/mirage/capnp-rpc>.