Remove compiler flags for flambda
This was needed to get flambda to inline the functor, but I don't think
anyone is using flambda and it massively slows down the build for
everyone else. We can try optimising it again once flambda2 is ready.

On OCaml 4.13 on my machine, this reduced the build time from 16s to 3s
(tested with `rm -rf _build && time dune build --release`).